requirements.txt
Django==4.1.2
djangorestframework==3.14.0
django-import-export==3.0.1
django-filter==22.1


